feat(Test):Development of the test application
Note: This test application covers below

	a. Retrieve the device’s full-scale configuration
	b. Retrieve and set the device’s output data rate
	c. Enable or disable the device’s interrupt pin
	d. Read the output data of a specific axis

+++ hal/i2c/i2c_hal.h
@@ -11,10 +11,10 @@
 #include "Board.h"
 #include "stdint.h"
 
-uint8_t i2c_read(uint8_t bus_address, uint8_t register_address, uint8_t *buffer,
+int32_t i2c_read(uint8_t bus_address, uint8_t register_address, uint8_t *buffer,
 uint16_t length);
 
-uint8_t i2c_write(uint8_t bus_address, uint8_t register_address, uint8_t *data,
+int32_t i2c_write(uint8_t bus_address, uint8_t register_address, uint8_t *data,
 uint16_t length);
 
 #endif
